Updated Sound Setting Logic
Changed the saved volume values from floats to integers.
This reduces the headache of dealing with floating point precision.
The change also allows for finer volume granularity to be implemented easily.

The settings script also received some simplification.

+++ Assets/Scripts/Settings/GameValues.cs
@@ -30,8 +30,8 @@ public class GameValues
     public float musicFadeOutFast;
     public float musicFadeOutSlow;
 
-    public float musicDefaultVolume;
-    public float soundEffectsDefaultVolume;
+    public int musicDefaultVolume;
+    public int soundEffectsDefaultVolume;
     public bool vibrationEnabledDefault;
     public bool foodSuitsEnabledDefault;
